modified

+++ src/GRAPH/AlienDictionary.java
@@ -62,7 +62,7 @@ public class AlienDictionary {
         String answer  = "";
 
         for(int it : topo){
-            answer = answer + (it + (int)('a'));
+            answer = answer + (it +'a');
         }
         return answer;
     }
